requirements.txt
numpy~=1.22.4
pytest==7.2.0
coverage==7.1.0
flake8==6.0.0
google~=3.0.0
PyYAML~=5.3.1